Reorganize introspection module
