+++ crates/apollo-mcp-server/src/introspection/tools.rs
@@ -1,0 +1,4 @@
+//! MCP tools to allow an AI agent to introspect a GraphQL schema and execute operations.
+
+pub(crate) mod execute;
+pub(crate) mod introspect;
